+++ src/main/java/seedu/duke/commands/project/ProjectSelectCommand.java
@@ -1,0 +1,30 @@
+package seedu.duke.commands.project;
+
+import seedu.duke.commands.Command;
+import seedu.duke.project.ProjectList;
+
+public class ProjectSelectCommand extends Command {
+
+    private int itemIndex;
+
+    public ProjectSelectCommand(int itemIndex) {
+        this.itemIndex = itemIndex;
+    }
+
+    public void executeCommand(ProjectList projects) {
+        try {
+            projects.selectProject(itemIndex);
+        } catch (NumberFormatException | IndexOutOfBoundsException | NullPointerException e) {
+            if (projects.getProjectList().size() == 0) {
+                System.out.println("I am empty!!!");
+            } else {
+                System.out.println("Invalid project ID");
+            }
+        }
+    }
+
+    public Boolean isExit() {
+        return false;
+    }
+
+}
