+++ src/main/java/seedu/duke/member/TeamMemberList.java
@@ -1,4 +1,4 @@
-package seedu.duke.memberStuff;
+package seedu.duke.member;
 
 import java.util.ArrayList;
